Programacao-Java-e-Desenvolvimento-agil/Modelagem-Agil/LeituraDeLivros/src/br/com/coursera/dominio/Livro.java
package br.com.coursera.dominio;

public class Livro {
	
	private String titulo;
	private String autor;
	private Estilo estilo;
	private int paginas;
	private Double pontosDeLeitura;

	public Livro(String titulo, String autor, Estilo estilo, int paginas) {
		this.titulo = titulo;
		this.autor = autor;
		this.estilo = estilo;
		this.paginas = paginas;
		this.pontosDeLeitura = 0.0;
	}

	public String getTitulo() {
		return titulo;
	}

	public void setTitulo(String titulo) {
		this.titulo = titulo;
	}

	public String getAutor() {
		return autor;
	}

	public void setAutor(String autor) {
		this.autor = autor;
	}

	public Estilo getEstilo() {
		return estilo;
	}

	public void setEstilo(Estilo estilo) {
		this.estilo = estilo;
	}

	public int getPaginas() {
		return paginas;
	}

	public void setPaginas(int paginas) {
		this.paginas = paginas;
	}
	
	public Double getPontosDeLeitura() {
		return pontosDeLeitura;
	}
	
	public void setPontosDeLeitura(Double pontosDeLeitura) {
		this.pontosDeLeitura = pontosDeLeitura;
	}

	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((autor == null) ? 0 : autor.hashCode());
		result = prime * result + ((estilo == null) ? 0 : estilo.hashCode());
		result = prime * result + paginas;
		result = prime * result + ((titulo == null) ? 0 : titulo.hashCode());
		return result;
	}

	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (obj == null)
			return false;
		if (getClass() != obj.getClass())
			return false;
		Livro other = (Livro) obj;
		if (autor == null) {
			if (other.autor != null)
				return false;
		} else if (!autor.equals(other.autor))
			return false;
		if (estilo != other.estilo)
			return false;
		if (paginas != other.paginas)
			return false;
		if (titulo == null) {
			if (other.titulo != null)
				return false;
		} else if (!titulo.equals(other.titulo))
			return false;
		return true;
	}

}
